エンティティ & リポジトリーのエラー解消 & 今週の達成リスト表示

+++ sanda-zero-carbon-app-backend/src/main/java/jp/kobespiral/sandazerocarbonappbackend/domain/repository/MapRepository.java
@@ -19,5 +19,5 @@ public interface MapRepository extends CrudRepository<Map, Long> {
      * @param level           レベル
      * @return Map マップ
      */
-    Map findByCurrentLocationAndLevel(Long currentLocation, int level);
+    Map findByCurrentLocationAndLevel(int currentLocation, int level);
 }
